moved-components-to-RecipeCollectionPage-folder
