+++ myCookingManager/frontend/src/components/RecipeCollectionPage/CategoryMenu.jsx
