+++ myCookingManager/frontend/src/components/RecipeCollectionPage/SearchBar.jsx
