+++ myCookingManager/frontend/src/components/global_components/App.jsx
@@ -4,7 +4,7 @@ import { useAuth0 } from "@auth0/auth0-react";
 import GlobalStyles from "./GlobalStyles";
 import Header from "./Header";
 import Home from "../HomePage/Home";
-import RecipeCollection from "../RecipeCollection";
+import RecipeCollection from "../RecipeCollectionPage/RecipeCollection";
 import RecipeDetails from "../RecipeDetails";
 import Callback from "./Callback";
 import CataloguingPage from "../CataloguingPage";
